feat(tmux): add tmux models, command builder, parser adapter, and legacy re-exports
Add the tmux model types, command builder, parser adapter, and turn the old tmux_commands / tmux_parser files into thin legacy re-exports. PaneNavigator is also moved to the new tmux_models import.

Changed files:
- lib/services/tmux/tmux_models.dart
- lib/services/tmux/tmux_command_builder.dart
- lib/services/tmux/tmux_parser_adapter.dart
- lib/services/tmux/tmux_commands.dart
- lib/services/tmux/tmux_parser.dart
- lib/services/tmux/pane_navigator.dart
